// src/app/Models/EmpRegister.ts
export interface EmpRegister {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    fatherName: string;
    gender: string;
    phone: string;
    created: Date;
    dateofBirth: Date;
    doa:Date;
    motherTongue: string;
    EmpId:number;
    standers: string;
    standerss: string[];
    classSec: string;
    classLevel: string;
    subjects:string;
    registrationNo: string;
    Nationality: string;
    medium: string; 
}
 